WIP: move _digitization inside channels
