+++ mne/channels/_digitization/__init__.py
