+++ mne/io/tests/test_meas_info.py
@@ -24,7 +24,7 @@ from mne.io.write import DATE_NONE
 from mne.io.meas_info import (Info, create_info, _merge_info,
                               _force_update_info, RAW_INFO_FIELDS,
                               _bad_chans_comp, _get_valid_units)
-from mne._digitization._utils import (_write_dig_points, _read_dig_points,
+from mne.channels._digitization._utils import (_write_dig_points, _read_dig_points,
                                       _make_dig_points,)
 from mne.io import read_raw_ctf
 from mne.utils import _TempDir, run_tests_if_main, catch_logging
